requirements.txt
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.8
django-axes==8.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-extensions==4.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
humanize==4.14.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
MarkupSafe==3.0.3
msgpack==1.1.2
packaging==25.0
pillow==12.0.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==7.1.0
requests==2.32.5
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tornado==6.5.2
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
Werkzeug==3.1.3
zope.interface==8.1.1
